Swagger generation as part of CI
