+++ ci/check/check.go
@@ -23,7 +23,7 @@ import (
 	"github.com/magefile/mage/mg"
 	"github.com/magefile/mage/sh"
 	"github.com/mysteriumnetwork/go-ci/commands"
-	"github.com/mysteriumnetwork/go-ci/util"
+	"github.com/mysteriumnetwork/node/ci/packages"
 )
 
 // Check performs commons checks.
@@ -54,31 +54,10 @@ func CheckGoImports() error {
 
 // CheckSwagger checks whether swagger spec at "tequilapi.json" is valid against swagger specification 2.0.
 func CheckSwagger() error {
-	mg.Deps(GetSwagger)
-
-	err := sh.RunV("swagger", "generate", "spec", "-o", "tequilapi.json", "--scan-models")
-	if err != nil {
-		fmt.Println("could not generate swagger spec")
-		return err
-	}
+	mg.Deps(packages.GenerateSwagger)
 
 	if err := sh.RunV("swagger", "validate", "tequilapi.json"); err != nil {
 		fmt.Println("could not validate swagger spec")
-		return err
-	}
-	return nil
-}
-
-// GetSwagger installs swagger tool.
-func GetSwagger() error {
-	path, _ := util.GetGoBinaryPath("swagger")
-	if path != "" {
-		fmt.Println("Tool 'swagger' already installed")
-		return nil
-	}
-	err := sh.RunV("go", "get", "-u", "github.com/go-swagger/go-swagger/cmd/swagger")
-	if err != nil {
-		fmt.Println("could not go get swagger")
 		return err
 	}
 	return nil
